Rename binary package -> files
